*: little refactor
add batch insert function
